Implemented part 2 of Day 13. Modified the lcm util function to accept a long array. Modified classes that use the lcm util function.

+++ src/main/java/aoc2019/Day12.java
@@ -162,7 +162,7 @@ public class Day12
         List<Integer> repeatFrequencies = new ArrayList<>();
         for( List<Integer> list : dimensions )
             repeatFrequencies.add( getIndexAtWhichCycleRepeats( list.stream().mapToInt( a -> a).toArray() ) );
-        return Util.lcm( repeatFrequencies.stream().mapToInt( a -> a ).toArray() );
+        return Util.lcm( repeatFrequencies.stream().mapToLong( a -> a ).toArray() );
 
     }
